+++ openvidu-call-react/src/components/chat/ChatHandsFree.css
@@ -81,7 +81,7 @@
     position: absolute;
     z-index: 99999;
     /* display: none; */
-    top: -8.7%;
+    top: 1%;
     /* left: ; */
     right: 3.8%;
     /* bottom: 0; */
@@ -100,7 +100,7 @@
 
   .isRecog{
     position:relative;
-    height: 50%;
+    height: 103%;
     /* left: -350%; */
     left: 3.7%;
     width: 92.7%;
@@ -302,3 +302,29 @@
     background-color: #eeeeee !important;
   }
   
+
+.mackdoong-switch{
+  position: absolute;
+  left: 5%;
+  /* display: none; */
+  font-size: 25px;
+}
+  
+
+.mackdoong-logo{
+  /* size: 40px; */
+  /* width: 40px; */
+  left: 23%;
+  position: absolute;
+  /* height: 40px; */
+  /* width: 30px */
+}
+
+
+.mackdoong-txt{
+  /* display: none; */
+  position: absolute;
+  bottom: 26%;
+  right: 7%;
+  width: 100px;
+}
