+++ openvidu-call-react/src/components/chat/ChatHandsFree.js
@@ -79,7 +79,6 @@ class ChatHandsFree extends PureComponent {
             startTime: data.startTime,
             marker: this.state.isStar,
             id: this.state.msgIndex,
-            play: false,
           });
           this.setState({
             msgIndex: this.state.msgIndex + 1,
@@ -142,7 +141,7 @@ class ChatHandsFree extends PureComponent {
       try {
         this.chatScroll.current.scrollTop =
           this.chatScroll.current.scrollHeight;
-      } catch (err) {}
+      } catch (err) { }
     }, 20);
   }
 
@@ -151,10 +150,11 @@ class ChatHandsFree extends PureComponent {
   };
 
   parentFunction = (data) => {
-    console.log("data = ", data);
-    console.log("아이디 식별자", this.props.localUser.getConnectionId());
-    this.state.message = data.text;
-    this.state.startTime = data.startTime;
+    console.log("text ==", data.text);
+    this.setState({
+      message: data.text,
+      startTime: data.startTime,
+    });
     if (
       data.text.includes("막둥아 기록 중지") ||
       data.text.includes("막둥아 기록중지") ||
@@ -207,56 +207,28 @@ class ChatHandsFree extends PureComponent {
         <div className='isRecog'>
           <div className='writingStatus'>
             <div
-              className={`inline-block vertical-align mr-20 ${
-                this.state.isRecog ? "colorYellow" : "colorRed"
-              }`}
+              className={`mackdoong-switch ${this.state.isRecog ? "colorYellow" : "colorRed"
+                }`}
             >
-              {this.state.isRecog ? "ON " : "OFF"}
+              {this.state.isRecog ? "ON" : "OFF"}
             </div>
-            <div className='inline-block vertical-align mr-8'>
+            <div className='mackdoong-logo'>
               <img
                 alt='막둥이'
                 src={this.state.isRecog ? isWriting : isNotWriting}
-                height={this.state.isRecog ? "40" : "20"}
-                width={this.state.isRecog ? "40" : "20"}
-                // size={100}
+                height={this.state.isRecog ? "90" : "90"}
+                width={this.state.isRecog ? "90" : "90"}
               />
             </div>
             <div
-              className='inline-block vertical-align'
+              className='mackdoong-txt'
               style={{ marginBottom: 4 }}
             >
               {this.state.isRecog
                 ? "막둥이가 기록 중이에요!"
-                : " 막둥이를 불러주세요!  "}
+                : "막둥이를 불러주세요!  "}
             </div>
           </div>
-
-          {/* <div className='writingStatus'>
-            <div
-              className={`inline-block vertical-align mr-20 ${!this.state.isRecog ? "colorYellow" : "colorRed"
-                }`}
-            >
-              {!this.state.isRecog ? "ON " : "OFF"}
-            </div>
-            <div className='inline-block vertical-align mr-8'>
-              <img
-                alt='막둥이'
-                src={!this.state.isRecog ? isWriting : isNotWriting}
-                height={!this.state.isRecog ? "60" : "60"}
-                width={!this.state.isRecog ? "60" : "360"}
-                // size={100}
-              />
-            </div>
-            <div
-              className='inline-block vertical-align'
-              style={{ marginBottom: 4 }}
-            >
-              {!this.state.isRecog
-                ? "막둥이가 기록 중이에요!"
-                : " 막둥이를 불러주세요!  "}
-            </div>
-          </div> */}
         </div>
 
         <div id='chatContainer'>
@@ -269,7 +241,7 @@ class ChatHandsFree extends PureComponent {
                   className={
                     "message" +
                     (data.connectionId !==
-                    this.props.localUser.getConnectionId()
+                      this.props.localUser.getConnectionId()
                       ? " left"
                       : " right")
                   }
@@ -284,12 +256,11 @@ class ChatHandsFree extends PureComponent {
 
                     <div
                       className={`
-                      ${
-                        data.connectionId !==
-                        this.props.localUser.getConnectionId()
+                      ${data.connectionId !==
+                          this.props.localUser.getConnectionId()
                           ? " f-left"
                           : " f-right"
-                      } ${data.marker ? "msg-content-star" : "msg-content"}
+                        } ${data.marker ? "msg-content-star" : "msg-content"}
                       `}
                     >
                       {/* <span className='triangle' /> */}
