+++ openvidu-call-react/src/components/edit/Temp.js
@@ -9,4 +9,3 @@ const Temp = () => {
 };
 
 export default Temp;
-
