+++ openvidu-call-react/src/components/edit/chat/Chatitem.css
@@ -22,12 +22,3 @@
 	background-color:blanchedalmond;
   color: black;
 }
-
-.chattime-buttons {
-  /* border-radius:4px; */
-  /* border: 1px solid whitesmoke; */
-  /* font-weight: bold; */
-  /* background-color: #e4ddd4; */
-  /* margin-left: 2px; */
-  /* color: #337bc4; */
-}
